% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump2sqlite.R
\name{dump2sqlite}
\alias{dump2sqlite}
\title{Converts MySQL dump to SQLite file}
\usage{
dump2sqlite(in_file, out_file = paste0(in_file, "ite"))
}
\arguments{
\item{in_file}{MySQL dump file path}

\item{out_file}{SQLite file path}
}
\description{
Converts MySQL dump to SQLite file
}
