% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_create.R
\name{get_create}
\alias{get_create}
\title{Extracts CREATE statement from MySQL dump}
\usage{
get_create(dump, table)
}
\arguments{
\item{dump}{A character vector containing lines of MySQL dump file}

\item{table}{Name of table to be created}
}
\value{
An SQL CREATE query
}
\description{
Extracts CREATE statement from MySQL dump
}
