% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tables}
\alias{get_tables}
\title{Extracts table names from MySQL dump}
\usage{
get_tables(dump)
}
\arguments{
\item{dump}{A character vector containing lines of MySQL dump file}
}
\value{
A character vector of table names
}
\description{
Extracts table names from MySQL dump
}
